6.0.7

+++ events/guild/messageReactionAdd.js
@@ -4,9 +4,12 @@ const Schema = require("../../databases/reaction-roles");
 const Schema2 = require("../../databases/reaction-roles2");
 
 module.exports = async (client, reaction, user) => {
-  let guild = client.guilds.cache.get(reaction.message.guildId)
-  let member = await guild.members.cache.get(user.id)
-  let channel = await guild.channels.cache.get(reaction.message.channelId)
+  let guild = client.guilds.cache.get(reaction.message.guildId);
+  let member = await guild.members.cache.get(user.id);
+  let channel = await guild.channels.cache.get(reaction.message.channelId);
+  let premium = await client.Premium.findOne({
+    GuildId: guild.id
+  });
   let ss = await client.Settings.findOne({ GuildId : guild.id });
   let es = ss.Embed;
   let ls = ss.Language;
@@ -47,39 +50,43 @@ module.exports = async (client, reaction, user) => {
   } catch (error) {
     console.log(error)
   }
-  try {
-    Schema2.findOne({
-      Message: reaction.message.id
-    }, async (err, data) => {
-      if (!data) return;
-      if (!Object.keys(data.RolesId).includes(reaction.emoji.name)) return;
+  
+  if (premium) {
+    try {
+      Schema2.findOne({
+        Message: reaction.message.id
+      }, async (err, data) => {
+        if (!data) return;
+        if (!Object.keys(data.RolesId).includes(reaction.emoji.name)) return;
 
-      const [roleid] = data.RolesId[reaction.emoji.name];
+        const [roleid] = data.RolesId[reaction.emoji.name];
 
-      try {
-        await member.roles.add(roleid)
-      } catch {
-        return channel.send(`${user} You already have that role!`)
-      }
-      // return channel.send(`${user} You have obtain the <@&${roleid}> role!`)
-      return channel.send({
-          embeds: [new MessageEmbed()
-            .setColor(es.color)
-            .setFooter(client.getFooter(es))
-            .setDescription(`${user} You have obtain the <@&${roleid}> role!`)
-          ]
-        })
-        .then((msg) => {
-          try {
-            setTimeout(() => {
-              msg.delete().catch(() => {});
-            }, 6000);
-          } catch (error) {
-            console.log(error)
-          }
-        })
-    });
-  } catch (error) {
-    console.log(error)
+        try {
+          await member.roles.add(roleid)
+        } catch {
+          return channel.send(`${user} You already have that role!`)
+        }
+        return channel.send({
+            embeds: [new MessageEmbed()
+              .setColor(es.color)
+              .setFooter(client.getFooter(es))
+              .setDescription(`${user} You have obtain the <@&${roleid}> role!`)
+            ]
+          })
+          .then((msg) => {
+            try {
+              setTimeout(() => {
+                msg.delete().catch(() => {});
+              }, 6000);
+            } catch (error) {
+              console.log(error)
+            }
+          })
+      });
+    } catch (error) {
+      console.log(error)
+    }
+  } else {
+    channel.send(`Your guild is **not premium now**, please upgrade to premium so you can activate Reaction Roles again`)
   }
 }
