+++ events/guild/messageReactionRemove.js
@@ -6,9 +6,12 @@ const SettingsSchema = require('../../databases/settings');
 const ee = require('../../botconfig/embed');
 
 module.exports = async (client, reaction, user) => {
-  let guild = client.guilds.cache.get(reaction.message.guildId)
-  let member = await guild.members.cache.get(user.id)
-  let channel = await guild.channels.cache.get(reaction.message.channelId)
+  let guild = client.guilds.cache.get(reaction.message.guildId);
+  let member = await guild.members.cache.get(user.id);
+  let premium = await client.Premium.findOne({
+    GuildId: guild.id
+  });
+  let channel = await guild.channels.cache.get(reaction.message.channelId);
   let ss = await client.Settings.findOne({ GuildId : guild.id });
   let es = ss.Embed;
   let ls = ss.Language;
@@ -48,37 +51,40 @@ module.exports = async (client, reaction, user) => {
   } catch (error) {
     console.log(error)
   }
-  try {
-    Schema2.findOne({
-      Message: reaction.message.id
-    }, async (err, data) => {
-      if (!data) return;
-      if (!Object.keys(data.RolesId).includes(reaction.emoji.name)) return;
+  if (premium) {
+    try {
+      Schema2.findOne({
+        Message: reaction.message.id
+      }, async (err, data) => {
+        if (!data) return;
+        if (!Object.keys(data.RolesId).includes(reaction.emoji.name)) return;
 
-      const [roleid] = data.RolesId[reaction.emoji.name];
-      try {
-        await member.roles.remove(roleid)
-      } catch {
-        return channel.send(`${user} You don't have that role!`)
-      }
-      // return channel.send(`${user} <@&${roleid}> role has been deleted!`)
-      return channel.send({
-        embeds: [new MessageEmbed()
-          .setColor(es.wrongcolor)
-          .setFooter(client.getFooter(es))
-          .setDescription(`${user} <@&${roleid}> role has been deleted!`)
-        ]
-      }).then((msg) => {
+        const [roleid] = data.RolesId[reaction.emoji.name];
         try {
-          setTimeout(() => {
-            msg.delete().catch(() => {});
-          }, 6000);
-        } catch (error) {
-          console.log(error)
+          await member.roles.remove(roleid)
+        } catch {
+          return channel.send(`${user} You don't have that role!`)
         }
-      })
-    });
-  } catch (error) {
-    console.log(error)
+        return channel.send({
+          embeds: [new MessageEmbed()
+            .setColor(es.wrongcolor)
+            .setFooter(client.getFooter(es))
+            .setDescription(`${user} <@&${roleid}> role has been deleted!`)
+          ]
+        }).then((msg) => {
+          try {
+            setTimeout(() => {
+              msg.delete().catch(() => {});
+            }, 6000);
+          } catch (error) {
+            console.log(error)
+          }
+        })
+      });
+    } catch (error) {
+      console.log(error)
+    }
+  } else {
+    channel.send(`Your guild is **not premium now**, please upgrade to premium so you can activate Reaction Roles again`)
   }
 }
